feat: User 도메인, 엔티티 이름 Account로 변경
