+++ charlee-web/src/main/java/com/efa73/charleeweb/account/interfaces/dto/LoginResponse.java
@@ -1,0 +1,9 @@
+package com.efa73.charleeweb.account.interfaces.dto;
+
+public record LoginResponse(
+        String email
+) {
+    public static LoginResponse of(String email) {
+        return new LoginResponse(email);
+    }
+}
