+++ charlee-web/src/main/java/com/efa73/charleeweb/account/login/filter/CustomAuthenticationFilter.java
@@ -1,0 +1,51 @@
+package com.efa73.charleeweb.account.login.filter;
+
+import com.efa73.charleeweb.account.interfaces.dto.LoginRequest;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import org.springframework.security.authentication.AuthenticationServiceException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
+
+public class CustomAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
+
+    private static final String DEFAULT_LOGIN_REQUEST_URL = "/api/login";
+    //    private static final String HTTP_METHOD = "POST";
+    private static final String CONTENT_TYPE = "application/json";
+//    private static final AntPathRequestMatcher DEFAULT_LOGIN_PATH_REQUEST_MATCHER =
+//            new AntPathRequestMatcher(DEFAULT_LOGIN_REQUEST_URL, HTTP_METHOD);
+
+    private final ObjectMapper objectMapper;
+
+    public CustomAuthenticationFilter(ObjectMapper objectMapper) {
+        super(DEFAULT_LOGIN_REQUEST_URL);
+        this.objectMapper = objectMapper;
+    }
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response)
+            throws AuthenticationException, IOException, ServletException {
+        if (request.getContentType() == null || !request.getContentType().equals(CONTENT_TYPE)) {
+            throw new AuthenticationServiceException(
+                    "Authentication Content-Type not supported: " + request.getContentType());
+        }
+
+        LoginRequest loginRequest = objectMapper.readValue(request.getInputStream(), LoginRequest.class);
+
+        UsernamePasswordAuthenticationToken authRequest = new UsernamePasswordAuthenticationToken(
+                loginRequest.email(),
+                loginRequest.password()
+        );
+
+        try {
+            return this.getAuthenticationManager().authenticate(authRequest);
+        } catch (AuthenticationException e) {
+            throw e;
+        }
+    }
+}
