+++ charlee-web/src/main/java/com/efa73/charleeweb/account/login/handler/LoginFailureHandler.java
@@ -1,0 +1,39 @@
+package com.efa73.charleeweb.account.login.handler;
+
+import com.efa73.charleeweb.common.dto.ExceptionResponse;
+import com.efa73.charleeweb.common.exception.CommonErrorCode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
+
+@Slf4j
+@RequiredArgsConstructor
+public class LoginFailureHandler extends SimpleUrlAuthenticationFailureHandler {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void onAuthenticationFailure(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            AuthenticationException exception) throws IOException, ServletException {
+        ResponseEntity<ExceptionResponse> errorResponse =
+                ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(ExceptionResponse.of(CommonErrorCode.LOGIN_FAILED));
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        response.setCharacterEncoding("UTF-8");
+        response.setContentType("application/json");
+        objectMapper.writeValue(
+                response.getWriter(),
+                errorResponse.getBody()
+        );
+    }
+}
