+++ charlee-web/src/main/java/com/efa73/charleeweb/account/login/service/CustomUserDetailsService.java
@@ -1,0 +1,29 @@
+package com.efa73.charleeweb.account.login.service;
+
+import com.efa73.charleeweb.account.domain.repository.UserRepository;
+import java.util.Collections;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return userRepository.findByEmail(email)
+                .map(user -> new User(
+                        user.getEmail(),
+                        user.getPassword(),
+                        Collections.singletonList(new SimpleGrantedAuthority(user.getRole().name()))
+                ))
+                .orElseThrow(() -> new UsernameNotFoundException("User not found with email: " + email));
+    }
+}
