--- charlee-web/src/main/java/com/efa73/charleeweb/user/interfaces/dto/LoginRequest.java
@@ -1,14 +1,0 @@
-package com.efa73.charleeweb.user.interfaces.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @Email(message = "Invalid email format")
-        @NotBlank(message = "Email cannot be blank")
-        String email,
-
-        @NotBlank(message = "Password cannot be blank")
-        String password
-) {
-}
