--- charlee-web/src/main/java/com/efa73/charleeweb/user/interfaces/dto/LoginResponse.java
@@ -1,9 +1,0 @@
-package com.efa73.charleeweb.user.interfaces.dto;
-
-public record LoginResponse(
-        String email
-) {
-    public static LoginResponse of(String email) {
-        return new LoginResponse(email);
-    }
-}
