+++ charlee-web/src/test/java/com/efa73/charleeweb/config/AdminInitializerTest.java
@@ -6,9 +6,9 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
-import com.efa73.charleeweb.user.domain.entity.Role;
-import com.efa73.charleeweb.user.domain.entity.User;
-import com.efa73.charleeweb.user.domain.repository.UserRepository;
+import com.efa73.charleeweb.account.domain.entity.Role;
+import com.efa73.charleeweb.account.domain.entity.Account;
+import com.efa73.charleeweb.account.domain.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -41,7 +41,7 @@ class AdminInitializerTest {
         }
 
         //then
-        verify(userRepository).save(any(User.class));
+        verify(userRepository).save(any(Account.class));
     }
 
     @Test
@@ -57,6 +57,6 @@ class AdminInitializerTest {
         }
 
         //then
-        verify(userRepository, never()).save(any(User.class));
+        verify(userRepository, never()).save(any(Account.class));
     }
 }
